#15 [feat] ItemTouchHelper.Callback 별도 클래스로 분리
